clean + isNothing to Cast (was Nested)
